Change move to trash side panel to be a modal
