+++ kolibri/plugins/facility/assets/src/constants.js
@@ -11,8 +11,6 @@ export const PageNames = {
   DATA_EXPORT_PAGE: 'DATA_EXPORT_PAGE',
   FACILITY_CONFIG_PAGE: 'FACILITY_CONFIG_PAGE',
   ALL_FACILITIES_PAGE: 'ALL_FACILITIES_PAGE',
-  MOVE_TO_TRASH_TRASH_SIDE_PANEL: 'MOVE_TO_TRASH_TRASH_SIDE_PANEL',
-  MOVE_TO_TRASH_TRASH_SIDE_PANEL__NEW_USERS: 'MOVE_TO_TRASH_TRASH_SIDE_PANEL__NEW_USERS',
   FILTER_USERS_SIDE_PANEL: 'FILTER_USERS_SIDE_PANEL',
   FILTER_USERS_SIDE_PANEL__NEW_USERS: 'FILTER_USERS_SIDE_PANEL__NEW_USERS',
   ADD_NEW_USER_SIDE_PANEL: 'ADD_NEW_USER_SIDE_PANEL',
